+++ kibot/misc.py
@@ -52,6 +52,7 @@ BLENDER_ERROR = 35
 WARN_AS_ERROR = 36
 CHECK_FIELD = 37
 IGNORED_ERRORS = 38
+GOT_WARNINGS = 39   # Not treated as errors, but using `--fail-on-warnings`
 error_level_to_name = ['NONE',
                        'INTERNAL_ERROR',
                        'WRONG_ARGUMENTS',
@@ -90,7 +91,8 @@ error_level_to_name = ['NONE',
                        'BLENDER_ERROR',
                        'WARN_AS_ERROR',
                        'CHECK_FIELD',
-                       'IGNORED_ERRORS'
+                       'IGNORED_ERRORS',
+                       'GOT_WARNINGS'
                        ]
 KICOST_SUBMODULE = '../submodules/KiCost/src/kicost'
 EXAMPLE_CFG = 'example_template.kibot.yaml'