+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- CLI:
+  - --fail-on-warnings: to return error when we detected warning (See #828)
 - iBoM: `mark_when_checked` option (#881)